Relocated and updated theme.js.
